+++ refnd/src/utils/mod.rs
@@ -3,5 +3,5 @@ pub mod fingerprints;
 mod read_sdf;
 
 pub use read_fasta::read_fasta;
-pub use fingerprints::{BitFingerprint, RealFingerprint};
+pub use fingerprints::{BitFingerprint, InlineBitSet, RealFingerprint};
 pub use read_sdf::{read_sdf, read_smi, read_csv_mol, read_molecule_file, FingerprintType};